Move chat interfaces into separate namespace
