+++ OutlookMatters.Core/Chat/IChatChannel.cs
@@ -1,6 +1,6 @@
 ﻿using OutlookMatters.Core.Settings;
 
-namespace OutlookMatters.Core.Mattermost.Interface
+namespace OutlookMatters.Core.Chat
 {
     public interface IChatChannel
     {
