+++ OutlookMatters.Core/ContextMenu/MailItemContextMenuEntry.cs
@@ -5,6 +5,7 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Error;
 using OutlookMatters.Core.Mail;
 using OutlookMatters.Core.Mattermost.Interface;
