+++ OutlookMatters.Core/Error/MessageBoxErrorDisplay.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Windows;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Error
