+++ OutlookMatters.Core/Mattermost/HttpImpl/ChatPostImpl.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Mattermost.HttpImpl
