+++ OutlookMatters.Core/Mattermost/HttpImpl/HttpSession.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Mattermost.HttpImpl
