+++ OutlookMatters.Core/Mattermost/HttpImpl/IChatChannelFactory.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Mattermost.HttpImpl
