+++ OutlookMatters.Core/Session/ISessionRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Session
